Move factory files to prisma/factories
